Optional mailing to localhost

+++ server.py
@@ -17,14 +17,28 @@ import os.path
 import shutil
 import subprocess
 import sys
+import yaml
+
+import mail_utils
 
 logging.basicConfig(level=logging.INFO)
 
 app = Flask(__name__)
 __request_num = 0
-REQUEST_DIR_PATH = 'request_dirs'
+SMTP_HOST = 'localhost'
+CONFIG_FILE = 'server_config.yaml'
+REQUEST_DIR_PATH = os.path.join(os.getcwd(), 'request_dirs')
 REQUEST_DIR_SEMAPHORE = gevent.lock.Semaphore()
 JOBS_DIR_PATH = os.path.join(os.getcwd(), 'jobs')
+
+
+def config():
+    if os.path.exists(CONFIG_FILE):
+        doc = yaml.load(open(CONFIG_FILE))
+        logging.info('config:\n%s' % open(CONFIG_FILE).read())
+        mail_utils.mail_filter = doc.get('mail_filter', '')
+    else:
+        logging.warning("No %s" % CONFIG_FILE)
 
 
 def request_path(request_num):
@@ -78,25 +92,24 @@ def run_command(request_dir, args):
     return stdoutdata, stderrdata
 
 
-@app.route("/")
-def hello():
-    return __name__
-
-
-@app.route("/jobs", methods=['GET', 'POST'])
-def jobs():
-    if request.method == 'GET':
-        return "TODO: show jobs"
-    elif request.method == 'POST':
-        ct = request.headers['Content-Type']
-        if ct == 'application/json':
-            return jobs_start(request.json)
-        else:
-            # unsupported media type
-            return Response(status=415)
-    else:
-        # method not allowed
-        return Response(405)
+def email_results(email_results_to, request_dir, job_name, stdoutdata, stderrdata):
+    if email_results_to:
+        # pack up the output and email
+        # if there are any return files, attach them
+        return_files_path = os.path.join(request_dir, 'return_files.txt')
+        attachments = []
+        if os.path.isfile(return_files_path):
+            for filename in open(return_files_path).readlines():
+                filepath = os.path.join(request_dir, filename.strip())
+                logging.info("Try attaching %s to email" % filepath)
+                if os.path.isfile(filepath):
+                    attachments.append(mail_utils.attachment_from_file(filepath))
+                else:
+                    logging.warning("%s is not a file" % filepath)
+        mail_utils.send_simple_mail(
+            SMTP_HOST, "job_server", email_results_to, "Results of %s" % job_name,
+            "stdout:\n%s\n\nstderr:\n%s\n\n" % (stdoutdata, stderrdata),
+            attachments = attachments)
 
 
 def jobs_start(request_json):
@@ -118,14 +131,38 @@ def jobs_start(request_json):
     request_num, request_dir = create_request_dir()
     logging.info("request %d: json %s" % (request_num, request_json))
     stdoutdata, stderrdata = do_job(request_num, request_dir, job_name)
+    email_results(request_json.get('email_results_to'), request_dir, job_name,
+                  stdoutdata, stderrdata)
     # remove request directory
-    shutil.rmtree(request_dir)
+    if not request_json.get('leave_output'):
+        shutil.rmtree(request_dir)
 
     # Respond
     resp = {'request_number': request_num,
             'stdout': stdoutdata,
             'stderr': stderrdata}
     return Response(json.dumps(resp), status=200, mimetype='application/json')
+
+
+@app.route("/")
+def hello():
+    return __name__
+
+
+@app.route("/jobs", methods=['GET', 'POST'])
+def jobs():
+    if request.method == 'GET':
+        return "TODO: show jobs"
+    elif request.method == 'POST':
+        ct = request.headers['Content-Type']
+        if ct == 'application/json':
+            return jobs_start(request.json)
+        else:
+            # unsupported media type
+            return Response(status=415)
+    else:
+        # method not allowed
+        return Response(405)
 
 
 def main():
@@ -137,6 +174,7 @@ def main():
         print >>sys.stderr, "Usage: %s [port]" % __file__
         sys.exit(1)
     logging.info("Running on port: %s" % port)
+    config()
     gevent.pywsgi.WSGIServer(("0.0.0.0", port), app).serve_forever()
